BindInt(64) require exact type.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/f070ffde7ce159aa8687e850b95c2a54f8caf5d3

commit f070ffde7ce159aa8687e850b95c2a54f8caf5d3
Author: Joshua Bell <jsbell@chromium.org>
Date:   Tue Dec 10 19:44:51 2019 +0000

    SQL: Disallow implicit integer downconversion on BindInt

    Calling sql::Statement::BindInt() with an int64_t argument silently
    downconverted, potentially saving incorrect data to a database.
    sql::Statement::BindInt64(int64_t) should be used instead.

    Introduce an uncallable sql::Statement::BindInt(int64_t) overload to
    fail the compile. This also catches calling BindInt with uint32_t. Fix
    up various callers that were silently downconverting.

    Bug: 908690

+++ components/brave_ads/browser/bundle_state_database.cc
@@ -341,9 +341,9 @@ bool BundleStateDatabase::InsertOrUpdateAdInfo(const ads::AdInfo& info) {
     ad_info_statement.BindString(5, info.end_timestamp);
     ad_info_statement.BindString(6, info.uuid);
     ad_info_statement.BindString(7, info.campaign_id);
-    ad_info_statement.BindInt(8, info.daily_cap);
-    ad_info_statement.BindInt(9, info.per_day);
-    ad_info_statement.BindInt(10, info.total_max);
+    ad_info_statement.BindInt(8, static_cast<int>(info.daily_cap));
+    ad_info_statement.BindInt(9, static_cast<int>(info.per_day));
+    ad_info_statement.BindInt(10, static_cast<int>(info.total_max));
     ad_info_statement.BindString(11, *it);
     if (!ad_info_statement.Run()) {
       return false;
@@ -373,7 +373,7 @@ bool BundleStateDatabase::InsertOrUpdateAdConversion(
   statement.BindString(0, ad_conversion.creative_set_id);
   statement.BindString(1, ad_conversion.type);
   statement.BindString(2, ad_conversion.url_pattern);
-  statement.BindInt(3, ad_conversion.observation_window);
+  statement.BindInt(3, static_cast<int>(ad_conversion.observation_window));
 
   return statement.Run();
 }
